+++ cogstat/cogstat_stat.py
@@ -1169,12 +1169,11 @@ def compare_groups_effect_size(pdf, dependent_var_name, groups, meas_level, samp
                 group_levels = sorted(set(pdf[groups + [dependent_var_name[0]]][groups[0]]))
                 if len(group_levels) == 2:
                     groups, grouped_data = _split_into_groups(pdf, dependent_var_name[0], groups)
-                    # convert pandas Float64 to float
                     pdf_result.loc[_("Cohen's d"), _('Value')] = \
-                        pingouin.compute_effsize(grouped_data[0].astype(float), grouped_data[1].astype(float),
+                        pingouin.compute_effsize(grouped_data[0], grouped_data[1],
                                                  paired=False, eftype='cohen')
                     pdf_result.loc[_("Eta-squared"), _('Value')] = \
-                        pingouin.compute_effsize(grouped_data[0].astype(float), grouped_data[1].astype(float),
+                        pingouin.compute_effsize(grouped_data[0], grouped_data[1],
                                                  paired=False, eftype='eta-square')
                 else:
                     standardized_effect_size_result = None
@@ -1204,8 +1203,7 @@ def compare_groups_effect_size(pdf, dependent_var_name, groups, meas_level, samp
                 group_levels = sorted(set(pdf[groups + [dependent_var_name[0]]][groups[0]]))
                 if len(group_levels) == 2:
                     groups, grouped_data = _split_into_groups(pdf, dependent_var_name[0], groups)
-                    # convert pandas Float64 to float
-                    hedges = pingouin.compute_effsize(grouped_data[0].astype(float), grouped_data[1].astype(float),
+                    hedges = pingouin.compute_effsize(grouped_data[0], grouped_data[1],
                                                       paired=False, eftype='hedges')
                     hedges_ci = pingouin.compute_esci(stat=hedges, nx=len(grouped_data[0]), ny=len(grouped_data[0]),
                                                       paired=False, eftype='cohen', confidence=0.95, decimals=3)
